refactor: player package 추가
