+++ src/main/kotlin/camp/nextstep/blackjack/game/BlackJackGame.kt
@@ -1,9 +1,9 @@
 package camp.nextstep.blackjack.game
 
 import camp.nextstep.blackjack.CardShuffler
-import camp.nextstep.blackjack.Player
 import camp.nextstep.blackjack.card.Card
 import camp.nextstep.blackjack.card.CardDeck
+import camp.nextstep.blackjack.player.Player
 
 class BlackJackGame private constructor(private var _cardDeck: CardDeck) {
